vpn_tray_app/requirements.txt
pystray>=0.19.4
Pillow>=9.5.0
websockets>=11.0.3 